+++ lib/ramda/internal/curried_method.rb
@@ -1,3 +1,5 @@
+require_relative 'function_with_arity'
+
 module Ramda
   module Internal
     # Curried Method
@@ -5,6 +7,7 @@ module Ramda
       def curried_method(name, &block)
         define_singleton_method(name) do |*args|
           curried = block.curry
+          # curried.define_singleton_method(:origin_arity) { block.arity }
 
           return curried if args.empty?