+++ lib/ramda/type.rb
@@ -5,6 +5,15 @@ module Ramda
   module Type
     extend ::Ramda::Internal::CurriedMethod
 
+    # See if an object (val) is an instance of the supplied constructor.
+    # This function will check up the inheritance chain, if any.
+    #
+    # (* -> {*}) -> a -> Boolean
+    #
+    curried_method(:is) do |type, x|
+      x.is_a?(type)
+    end
+
     # Checks if the input value is nil.
     curried_method(:is_nil, &:nil?)
   end