+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Not Released
 * ap
 * complement
 * empty - uses x.class.new for defining new empty object
+* from_pairs
 
 Release 0.2.3
 ---------------