+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Not Released
 * complement
 * empty - uses x.class.new for defining new empty object
 * from_pairs
+* is
 
 Release 0.2.3
 ---------------