+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Not Released
 * empty - uses x.class.new for defining new empty object
 * from_pairs
 * is
+* of
 
 Release 0.2.3
 ---------------